backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.1
PyJWT[crypto]>=2.8.0
svix>=1.21.0
supabase>=2.0.0
